requirements.txt
flask==3.0.0
nltk==3.8.1
requests==2.31.0
langchain==0.1.11
langchain-groq==0.0.1
python-dotenv==1.0.1
